Rename Performance tests to load tests.
