+++ v1/src/test/java/com/google/cloud/teleport/templates/PubsubToPubsubLT.java
@@ -13,30 +13,24 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.cloud.teleport.v2.templates.pubsubtotext;
+package com.google.cloud.teleport.templates;
 
-import static com.google.cloud.teleport.it.artifacts.ArtifactUtils.createGcsClient;
-import static com.google.cloud.teleport.it.artifacts.ArtifactUtils.getFullGcsPath;
 import static com.google.cloud.teleport.it.matchers.TemplateAsserts.assertThatPipeline;
 import static com.google.cloud.teleport.it.matchers.TemplateAsserts.assertThatResult;
 import static com.google.common.truth.Truth.assertThat;
 
-import com.google.cloud.storage.Storage;
 import com.google.cloud.teleport.it.DataGenerator;
-import com.google.cloud.teleport.it.PerformanceBenchmarkingBase;
+import com.google.cloud.teleport.it.TemplateLoadTestBase;
 import com.google.cloud.teleport.it.TestProperties;
-import com.google.cloud.teleport.it.artifacts.ArtifactClient;
-import com.google.cloud.teleport.it.artifacts.GcsArtifactClient;
 import com.google.cloud.teleport.it.launcher.PipelineLauncher.LaunchConfig;
 import com.google.cloud.teleport.it.launcher.PipelineLauncher.LaunchInfo;
 import com.google.cloud.teleport.it.launcher.PipelineOperator.Result;
 import com.google.cloud.teleport.it.pubsub.DefaultPubsubResourceManager;
 import com.google.cloud.teleport.it.pubsub.PubsubResourceManager;
-import com.google.cloud.teleport.metadata.TemplateIntegrationTest;
+import com.google.cloud.teleport.metadata.TemplateLoadTest;
 import com.google.common.base.MoreObjects;
 import com.google.pubsub.v1.SubscriptionName;
 import com.google.pubsub.v1.TopicName;
-import com.google.re2j.Pattern;
 import java.io.IOException;
 import java.text.ParseException;
 import java.time.Duration;
@@ -47,87 +41,67 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Performance tests for {@link PubsubToText PubSub to GCS Text} Flex template. */
-@TemplateIntegrationTest(PubsubToText.class)
+/** Performance tests for {@link PubsubToPubsub PubSub to PubSub} template. */
+@TemplateLoadTest(PubsubToPubsub.class)
 @RunWith(JUnit4.class)
-public final class PubsubToTextPerformanceIT extends PerformanceBenchmarkingBase {
-
-  private static final String ARTIFACT_BUCKET = TestProperties.artifactBucket();
+public class PubsubToPubsubLT extends TemplateLoadTestBase {
   private static final String SPEC_PATH =
       MoreObjects.firstNonNull(
-          TestProperties.specPath(),
-          "gs://dataflow-templates/latest/flex/Cloud_PubSub_to_GCS_Text_Flex");
-  private static final String TEST_ROOT_DIR =
-      PubsubToTextPerformanceIT.class.getSimpleName().toLowerCase();
-  private static final String INPUT_SUBSCRIPTION = "inputSubscription";
-  private static final String NUM_SHARDS_KEY = "numShards";
-  private static final String OUTPUT_DIRECTORY_KEY = "outputDirectory";
-  private static final String WINDOW_DURATION_KEY = "windowDuration";
-  private static final String NUM_WORKERS_KEY = "numWorkers";
-  private static final String MAX_WORKERS_KEY = "maxWorkers";
-  private static final String OUTPUT_FILENAME_PREFIX = "outputFilenamePrefix";
-  private static final String DEFAULT_WINDOW_DURATION = "10s";
-  private static final Pattern EXPECTED_PATTERN = Pattern.compile(".*subscription-output-.*");
+          TestProperties.specPath(), "gs://dataflow-templates/latest/Cloud_PubSub_to_Cloud_PubSub");
+  private static final long NUM_MESSAGES = 35000000L;
   private static final String INPUT_PCOLLECTION = "Read PubSub Events/PubsubUnboundedSource.out0";
-  private static final String OUTPUT_PCOLLECTION =
-      "Write File(s)/WriteFiles/WriteShardedBundlesToTempFiles/ApplyShardingKey.out0";
-  private static final long numMessages = 35000000L;
-
+  private static final String OUTPUT_PCOLLECTION = "Write PubSub Events/MapElements/Map.out0";
   private static PubsubResourceManager pubsubResourceManager;
-  private static ArtifactClient artifactClient;
 
   @Before
   public void setup() throws IOException {
-    // Set up resource managers
     pubsubResourceManager =
         DefaultPubsubResourceManager.builder(testName.getMethodName(), PROJECT)
             .credentialsProvider(CREDENTIALS_PROVIDER)
             .build();
-    Storage gcsClient = createGcsClient(CREDENTIALS);
-    artifactClient = GcsArtifactClient.builder(gcsClient, ARTIFACT_BUCKET, TEST_ROOT_DIR).build();
   }
 
   @After
-  public void tearDown() {
-    artifactClient.cleanupRun();
+  public void teardown() {
     pubsubResourceManager.cleanupAll();
   }
 
   @Test
-  public void testBacklog10gb() throws IOException, InterruptedException, ParseException {
+  public void testBacklog10gb() throws IOException, ParseException, InterruptedException {
     testBacklog10gb(Function.identity());
   }
 
   @Test
   public void testBacklog10gbUsingStreamingEngine()
-      throws IOException, InterruptedException, ParseException {
+      throws IOException, ParseException, InterruptedException {
     testBacklog10gb(config -> config.addEnvironment("enableStreamingEngine", true));
   }
 
   @Test
-  public void testSteadyState1hr() throws IOException, InterruptedException, ParseException {
+  public void testSteadyState1hr() throws IOException, ParseException, InterruptedException {
     testSteadyState1hr(Function.identity());
   }
 
   @Test
   public void testSteadyState1hrUsingStreamingEngine()
-      throws IOException, InterruptedException, ParseException {
+      throws IOException, ParseException, InterruptedException {
     testSteadyState1hr(config -> config.addEnvironment("enableStreamingEngine", true));
   }
 
   public void testBacklog10gb(Function<LaunchConfig.Builder, LaunchConfig.Builder> paramsAdder)
-      throws IOException, InterruptedException, ParseException {
+      throws IOException, ParseException, InterruptedException {
     // Arrange
-    String name = testName.getMethodName();
-    // Create topic and subscription
     TopicName backlogTopic = pubsubResourceManager.createTopic("backlog-input");
     SubscriptionName backlogSubscription =
         pubsubResourceManager.createSubscription(backlogTopic, "backlog-subscription");
+    SubscriptionName outputSubscription =
+        pubsubResourceManager.createSubscription(backlogTopic, "output-subscription");
+    TopicName outputTopic = pubsubResourceManager.createTopic("output");
     // Generate fake data to topic
     DataGenerator dataGenerator =
         DataGenerator.builderWithSchemaTemplate(testName, "GAME_EVENT")
             .setQPS("1000000")
-            .setMessagesLimit(String.valueOf(numMessages))
+            .setMessagesLimit(String.valueOf(NUM_MESSAGES))
             .setTopic(backlogTopic.toString())
             .setNumWorkers("50")
             .setMaxNumWorkers("100")
@@ -137,12 +111,9 @@ public final class PubsubToTextPerformanceIT extends PerformanceBenchmarkingBase
         paramsAdder
             .apply(
                 LaunchConfig.builder(testName, SPEC_PATH)
-                    .addParameter(INPUT_SUBSCRIPTION, backlogSubscription.toString())
-                    .addParameter(WINDOW_DURATION_KEY, DEFAULT_WINDOW_DURATION)
-                    .addParameter(OUTPUT_DIRECTORY_KEY, getTestMethodDirPath())
-                    .addParameter(NUM_SHARDS_KEY, "20")
-                    .addParameter(OUTPUT_FILENAME_PREFIX, "subscription-output-")
-                    .addParameter(NUM_WORKERS_KEY, "20"))
+                    .addEnvironment("maxWorkers", 100)
+                    .addParameter("inputSubscription", backlogSubscription.toString())
+                    .addParameter("outputTopic", outputTopic.toString()))
             .build();
 
     // Act
@@ -150,28 +121,32 @@ public final class PubsubToTextPerformanceIT extends PerformanceBenchmarkingBase
     assertThatPipeline(info).isRunning();
     Result result =
         pipelineOperator.waitForConditionAndFinish(
-            createConfig(info, Duration.ofMinutes(30)),
-            () -> waitForNumMessages(info.jobId(), INPUT_PCOLLECTION, numMessages));
+            createConfig(info, Duration.ofMinutes(60)),
+            () -> waitForNumMessages(info.jobId(), INPUT_PCOLLECTION, NUM_MESSAGES));
 
     // Assert
     assertThatResult(result).meetsConditions();
-    assertThat(artifactClient.listArtifacts(name, EXPECTED_PATTERN)).isNotEmpty();
+    // check to see if messages reached the output topic
+    assertThat(pubsubResourceManager.pull(outputSubscription, 5).getReceivedMessagesCount())
+        .isGreaterThan(0);
 
     // export results
     exportMetricsToBigQuery(info, getMetrics(info, INPUT_PCOLLECTION, OUTPUT_PCOLLECTION));
   }
 
   public void testSteadyState1hr(Function<LaunchConfig.Builder, LaunchConfig.Builder> paramsAdder)
-      throws IOException, InterruptedException, ParseException {
+      throws IOException, ParseException, InterruptedException {
     // Arrange
-    String name = testName.getMethodName();
-    TopicName topic = pubsubResourceManager.createTopic(testName.getMethodName() + "input");
-    SubscriptionName subscription =
-        pubsubResourceManager.createSubscription(topic, "input-subscription");
+    TopicName inputTopic = pubsubResourceManager.createTopic("steady-state-input");
+    SubscriptionName backlogSubscription =
+        pubsubResourceManager.createSubscription(inputTopic, "steady-state-subscription");
+    TopicName outputTopic = pubsubResourceManager.createTopic("output");
+    SubscriptionName outputSubscription =
+        pubsubResourceManager.createSubscription(inputTopic, "output-subscription");
     DataGenerator dataGenerator =
         DataGenerator.builderWithSchemaTemplate(testName, "GAME_EVENT")
             .setQPS("100000")
-            .setTopic(topic.toString())
+            .setTopic(inputTopic.toString())
             .setNumWorkers("10")
             .setMaxNumWorkers("100")
             .build();
@@ -179,12 +154,9 @@ public final class PubsubToTextPerformanceIT extends PerformanceBenchmarkingBase
         paramsAdder
             .apply(
                 LaunchConfig.builder(testName, SPEC_PATH)
-                    .addParameter(INPUT_SUBSCRIPTION, subscription.toString())
-                    .addParameter(WINDOW_DURATION_KEY, DEFAULT_WINDOW_DURATION)
-                    .addParameter(OUTPUT_DIRECTORY_KEY, getTestMethodDirPath())
-                    .addParameter(NUM_SHARDS_KEY, "20")
-                    .addParameter(OUTPUT_FILENAME_PREFIX, "subscription-output-")
-                    .addParameter(NUM_WORKERS_KEY, "10"))
+                    .addEnvironment("maxWorkers", 100)
+                    .addParameter("inputSubscription", backlogSubscription.toString())
+                    .addParameter("outputTopic", outputTopic.toString()))
             .build();
 
     // Act
@@ -192,17 +164,13 @@ public final class PubsubToTextPerformanceIT extends PerformanceBenchmarkingBase
     assertThatPipeline(info).isRunning();
     dataGenerator.execute(Duration.ofMinutes(60));
     Result result = pipelineOperator.drainJobAndFinish(createConfig(info, Duration.ofMinutes(20)));
-
     // Assert
     assertThat(result).isEqualTo(Result.LAUNCH_FINISHED);
-    assertThat(artifactClient.listArtifacts(name, EXPECTED_PATTERN)).isNotEmpty();
+    // check to see if messages reached the output topic
+    assertThat(pubsubResourceManager.pull(outputSubscription, 5).getReceivedMessagesCount())
+        .isGreaterThan(0);
 
     // export results
     exportMetricsToBigQuery(info, getMetrics(info, INPUT_PCOLLECTION, OUTPUT_PCOLLECTION));
-  }
-
-  private String getTestMethodDirPath() {
-    return getFullGcsPath(
-        ARTIFACT_BUCKET, TEST_ROOT_DIR, artifactClient.runId(), testName.getMethodName());
   }
 }
